// Returns-ASP-project/ClientApp/src/app/api/models/return-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ReturnDto {
  batchDate?: null | string;
  comment?: null | string;
  customer?: null | string;
  docDate?: null | string;
  docNo?: null | string;
  fault?: null | string;
  id?: null | string;
  owner?: null | string;
  product?: null | string;
  qtyOnDoc?: null | number;
  qtyReturned?: null | number;
  resolved?: null | boolean;
}